pyproject.toml
[project]
name = "restaurantflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "phonenumbers (>=9.0.5,<10.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "fastapi-debug-toolbar (>=0.6.3,<0.7.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "restaurantflow", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
